shim: fixed-width span lengths, and let gen-ffi see an out-of-tree shim
Two changes that cost nothing natively and are needed to relink this tier
for a 32-bit target.

Every FFI entry point taking a (const uint8_t*, len) span declared that
length as size_t. size_t is 64-bit on all four native targets, so nothing
changed there, but it is 32-bit on wasm32 -- and a wasm build of the same
shim then expects an i32 where the caller passes an i64. That mismatch
links with a warning and reads out of bounds at runtime, which is a far
worse failure than a link error. uint32_t is explicit and correct
everywhere.

gen-ffi's span collapser matched the literal "size_t", so it now accepts
either spelling; size_t still works for a shim written the old way.

SG_EXTRA_SHIM lets gen-ffi parse C signatures from sources outside this
repo. A consumer that relinks the canvas tier against a different host has
shim files of its own, and without this the manifest refuses to emit with
"no C signature found" for symbols that do exist.

Full suite green: canvas 55/55, input 89/89, frustum 32/32, phase9 11/11.

+++ shim/sg_input.cpp
@@ -278,7 +278,7 @@ extern "C" uint32_t sg_pad_has_rumble(uint32_t slot) {
 }
 
 /** Adds an SDL controller-mapping string. Returns 1 if new, 0 if updated. */
-extern "C" int32_t sg_pad_add_mapping(const uint8_t* text, size_t len) {
+extern "C" int32_t sg_pad_add_mapping(const uint8_t* text, uint32_t len) {
   char buf[1024];
   if (len >= sizeof(buf)) { sg_mail_set("mapping string too long"); return SG_ERANGE; }
   memcpy(buf, text, len);
